requirements.txt
agno == 1.5.10
dotenv == 0.9.9
duckduckgo-search == 8.0.2
fastapi == 0.115.12
fastf1 == 3.5.3
flask == 3.1.1
googlesearch-python == 1.3.0
mcp == 1.9.3
openai == 1.85.0
polars == 1.30.0
pycountry == 24.6.1
sqlalchemy == 2.0.41
wikipedia == 1.4.0
discord-py == 2.5.2